+++ trainer/trainlib/batch_loader.py
@@ -10,7 +10,7 @@ class BatchLoader:
     def __init__(self, path: str, seed: int, batch_size: int):
         self._loader = native.Loader(path, seed)
         self._points = numpy.zeros((batch_size, native.point_dims()), dtype=ctypes.c_float)
-        self._embeddings = numpy.zeros(batch_size, dtype=ctypes.c_int)
+        self._embeddings = numpy.zeros(batch_size, dtype=ctypes.c_float)
         self._targets = numpy.zeros((batch_size, 1), dtype=ctypes.c_float)
 
     def load_batch(self, device: torch.device) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
